Будажапова Екатерина. Задача 2. Вариант 15. Ленточная горизонтальная схема (#458)
### Sequential Version:
We check the input data and initialize the necessary variables. The
elements of matrix A and vector b are extracted from the input data, and
the dimensions of the matrix are calculated: the number of columns and
rows.
We validate the input data: ensuring that the number of rows in the
matrix and the number of columns in the vector are correct, and that the
matrix can be multiplied by the vector (i.e., the number of columns in
the matrix must match the number of rows in the vector).
The main logic for multiplying the matrix by the vector is implemented.
For each row of matrix A, its elements are multiplied by the
corresponding elements of vector b, and the result is stored in res.
The results of the multiplication are passed back through the output
data.

### Parallel Version:
The process (rank 0) reads matrix A and vector b. 
Values are sent from the root process to all other processes. 
We compare the number of rows and the number of processes, and based on
this, we distribute the rows of the matrix among all processes.
We calculate the number of rows to be sent and the offsets for each
process.
Similar to the sequential version, we perform validation. 
Each process multiplies its portion of the matrix local_A by vector b.
The results of the computation are collected in the vector local_res.
Then, the results from each process are gathered on the master process
using the gatherv function to combine the results into the final result
vector res.
On the master process, the final results are written to the output data.