Adjustable logging verbosity within non-root processes (#451)
Many often forget about what checks in tests can be done in non-root
workers, which leads to their exit with a non-zero return code and an
overall workflow fail due to the MPI runner panic. Sometimes it is not
so obvious, you can forget it and rack your brains for a long time,
looking for the culprit in a pile of tests.

In such cases, I advise commenting out the code responsible for
silencing gtest output in workers and running each worker in a separate
terminal window via mpiexec to understand what exactly causes the
problem. This is inconvenient, so this patch adds failed tests logging
in workers and opt-in for full logging via a special startup argument
(`--full-workers-log`).

Co-authored-by: Arseniy Obolenskiy <gooddoog@student.su>
Co-authored-by: Nesterov Alexander <nesterov.alexander@outlook.com>