Зайцев Артем. Задача 3. Вариант 25. Построение выпуклой оболочки. Проход Джарвиса. (#628)
## Jarvis march
1. Find the point, which is definitely is a part of the convex hull. It
is accepted to choose the rightmost of the points having the minimal
ordinate.
2. Among the others points it is necessary to find the most
counterclockwise one. It means that slack vector formed by such a point
and previously found start point has the minimal possible angle with all
the other vectors that can be formed by other points of the set and
selected the most counterclockwise one.
3. Repeat until the next point is the start point selected in p. 1.

## Sequential
1. Find the start point of the set, name it `start`.
2. Suggest that the next point of the given set is the next vertex of
convex hull. Name it `next`.
5. For all the points of the set that have never been used before in
this loop (name it `supposed`), check is triplet `(start, supposed,
next)` the most counterclockwise among the others.
6. If it does, let `next` be `supposed`. It's the next vertex of convex
hull.
7. Let `prev` be `next`, let `next` be the next point after actual
`next`, and repeat 2-5 while points `next` and `start` are differentent.

## Parallel
The same as ***Sequentional***, but checking of triplet being the most
counterclockwise is performed on several procs.

## struct Point
The struct defined in `seq/.../include/point.hpp` provides memebers `T
x, y`, which represents the coords of the Point. `operator*(const Point&
other)` represents vector product of vectors `(x, y)` and `(other.x,
other.y)`. It's needed to simplify and shorten the solution. Also
`b.between(const Point& a, const Point& c)` method is defined to check
whether point `b` belongs to `ac` segment. To be serializable by `boost`
library and be sendable and recieveable by `boost::mpi`,
`serialize(Archieve& a, const unsigned int version)` is overloaded for
`Point<T>` as specified in `boost` docs.